Cleanup and document compilation files
