+++ Source/Hotload/Compilation/CompileResult.cs
@@ -1,13 +1,31 @@
-﻿namespace Mocha.Hotload.Compilation;
+﻿namespace Mocha.Hotload;
 
-public struct CompileResult
+/// <summary>
+/// Represents a final compilation result.
+/// </summary>
+public readonly struct CompileResult
 {
-	public readonly bool WasSuccessful;
+	/// <summary>
+	/// Whether or not the compilation completed successfully.
+	/// </summary>
+	public bool WasSuccessful { get; }
 
-	public readonly byte[]? CompiledAssembly;
-	public readonly byte[]? CompiledAssemblySymbols;
-	public readonly string[]? Errors;
+	/// <summary>
+	/// The bytes of the compiled assembly.
+	/// </summary>
+	public byte[]? CompiledAssembly { get; }
+	/// <summary>
+	/// The bytes of the symbols in the assembly.
+	/// </summary>
+	public byte[]? CompiledAssemblySymbols { get; }
+	/// <summary>
+	/// An array of errors that occurred during compilation.
+	/// </summary>
+	public string[]? Errors { get; }
 
+	/// <summary>
+	/// Whether or not the compilation has assembly symbols.
+	/// </summary>
 	public bool HasSymbols => CompiledAssemblySymbols is not null;
 
 	private CompileResult( bool wasSuccessful, byte[]? compiledAssembly = null, byte[]? compiledAssemblySymbols = null, string[]? errors = null )
@@ -19,6 +37,11 @@ public struct CompileResult
 		Errors = errors;
 	}
 
+	/// <summary>
+	/// Shorthand method to create a failed <see cref="CompileResult"/>.
+	/// </summary>
+	/// <param name="errors">An array containing all of the errors that happened during the compilation.</param>
+	/// <returns>The newly created <see cref="CompileResult"/>.</returns>
 	public static CompileResult Failed( string[] errors )
 	{
 		return new CompileResult(
@@ -27,6 +50,12 @@ public struct CompileResult
 		);
 	}
 
+	/// <summary>
+	/// Shorthand method to create a successful <see cref="CompileResult"/>.
+	/// </summary>
+	/// <param name="compiledAssembly">The bytes of the compiled assembly.</param>
+	/// <param name="compiledAssemblySymbols">The bytes of the symbols contained in the compiled assembly.</param>
+	/// <returns>The newly created <see cref="CompileResult"/>.</returns>
 	public static CompileResult Successful( byte[] compiledAssembly, byte[]? compiledAssemblySymbols )
 	{
 		return new CompileResult(
